# notes-service/src/main/resources/static/openapi.yaml
# https://swagger.io/docs/specification/v3_0/basic-structure/
openapi: 3.0.4
info:
    title: Patient Notes API
    description: API for managing patient notes
      [access the GitHub repository here](https://github.com/The-Great-Toad/diabetes-risk-app/tree/main/notes-service)
    version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
tags:
    - name: Notes
      description: Operations related to patient notes

paths:
  /api/notes/{patientId}:
    get:
      security:
        - basicAuth: []
      responses:
        '200':
          description: A list of notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /notes:
    post:
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patientId:
                  type: string
                  example: 12345
                note:
                  type: string
                  example: This is a new note.
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

components:
  #  https://swagger.io/docs/specification/v3_0/authentication/basic-authentication/
  securitySchemes:
    basicAuth: # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
    Note:
      description: A note associated with a patient
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the note
          example: 1
        patientId:
          type: string
          description: Unique identifier for the patient
          example: 12345
        patient:
          type: string
          description: The name of the patient
          example: John Doe
        date:
          type: string
          description: The date of the note
          format: local-date
          example: 2025-04-18
        note:
          type: string
          description: The content of the note
          example: This is a new note.

security:
  - basicAuth: [] # <-- use the same name here